+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bug Fixes
 
 - fix default config settings for `ADMIN_LOGIN_ENDPOINT` and `ADMIN_LOGOUT_ENDPOINT`
+- defer initialization of the Admin extension to fix template overriding
 
 ### Breaking Changes